% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_boxes}
\alias{bootstrap_boxes}
\title{bBootstrap bounding boxes}
\usage{
bootstrap_boxes(x, scale = 2, size = nrow(x) * 10,
  quantiles = c(0.05, 0.95))
}
\arguments{
\item{x}{input 2-d matrix: first column is wisth, second column is height of the bounding box}

\item{scale}{how to rescale each box. For example 2 means it will randomly dowscaled-upscaled
within a range of 1/2 to 2. Which means it's size will be within a range being 2x smaller to 2x larger
then original bounding box}

\item{size}{size of bootstrapped result}

\item{quantiles}{keep resulting samples within these quantiles of the original distribution}
}
\description{
bootstrap bounding boxes in order to produce smooth box distribution
}
