% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{kmeans_iou}
\alias{kmeans_iou}
\title{K-means clustering of 2-d data with IoU distance}
\usage{
kmeans_iou(x, k, iter = 10, init = x[sample(nrow(x), size = k), , drop
  = FALSE], verbose = FALSE)
}
\arguments{
\item{x}{input 2-d matrix: first column is wisth, second column is height of the bounding box}

\item{k}{number of clusters}

\item{iter}{number of k-means iterations}

\item{init}{2-d matrix of initial centroids. Should have \code{k} rows}

\item{verbose}{logical, whether to print information during fitting the model}
}
\value{
list of 2 values. First is a matrix of cluster centers,
second is cluster membership indices.
}
\description{
Allows to cluster 2-dimensional with k-means using intersection over union distance.
This is useful for anchors initialization in the YOLO-family of models for object detection.
See Redmon et al. (2016) <arXiv:1612.08242>, Redmon et al. (2016) <arXiv:1804.02767>.
}
